backend/rag_service/requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
requests==2.32.3
openai==1.54.4
chromadb==0.5.23
numpy==1.26.4
llama-index-core==0.11.20
llama-index-vector-stores-chroma==0.2.0
llama-index-readers-file==0.2.2
llama-index-embeddings-huggingface==0.3.1
sentence-transformers==3.2.1
pypdf==4.3.1
python-pptx==1.0.2
